+++ reader.go
@@ -1,9 +1,10 @@
 package ishell
 
 import (
+	"bytes"
 	"sync"
 
-	"github.com/chzyer/readline"
+	"gopkg.in/readline.v1"
 )
 
 type (
@@ -13,14 +14,41 @@ type (
 	}
 
 	shellReader struct {
-		scanner   *readline.Instance
-		consumers []chan lineString
-		reading   bool
+		scanner      *readline.Instance
+		consumers    []chan lineString
+		reading      bool
+		readingMulti bool
+		buf          *bytes.Buffer
+		prompt       string
+		multiPrompt  string
+		showPrompt   bool
+		completer    readline.AutoCompleter
 		sync.Mutex
 	}
 )
 
-func (s *shellReader) ReadLine(consumer chan lineString) {
+// rlPrompt returns the proper prompt for readline based on showPrompt and
+// prompt members.
+func (s *shellReader) rlPrompt() string {
+	if s.showPrompt {
+		if s.readingMulti {
+			return s.multiPrompt
+		}
+		return s.prompt
+	}
+	return ""
+}
+
+func (s *shellReader) readPassword() string {
+	password, _ := s.scanner.ReadPassword("")
+	return string(password)
+}
+
+func (s *shellReader) setMultiMode(use bool) {
+	s.readingMulti = use
+}
+
+func (s *shellReader) readLine(consumer chan lineString) {
 	s.Lock()
 	defer s.Unlock()
 	s.consumers = append(s.consumers, consumer)
@@ -31,8 +59,25 @@ func (s *shellReader) ReadLine(consumer chan lineString) {
 	s.reading = true
 	// start reading
 	go func() {
+		// detect if print is called to
+		// prevent readline lib from clearing line.
+		// TODO find better way.
+		shellPrompt := s.prompt
+		prompt := s.rlPrompt()
+		if s.buf.Len() > 0 {
+			prompt += s.buf.String()
+			s.buf.Truncate(0)
+		}
+
+		// use printed statement as prompt
+		s.scanner.SetPrompt(prompt)
+
 		line, err := s.scanner.Readline()
-		ls := lineString{line, err}
+
+		// reset prompt
+		s.scanner.SetPrompt(shellPrompt)
+
+		ls := lineString{string(line), err}
 		s.Lock()
 		defer s.Unlock()
 		for i := range s.consumers {